rename OrderInfo to Checkout and /process route to /checkout
