+++ client/components/index.js
@@ -11,5 +11,5 @@ export { default as BrowseProducts } from './BrowseProducts';
 export { default as AllProducts } from './AllProducts';
 export { default as SingleProduct } from './SingleProduct';
 export { default as Cart } from './Cart';
-export { default as OrderInfo } from './OrderInfo';
+export { default as Checkout } from './Checkout';
 export { default as AllOrders } from './AllOrders';
