+++ client/utils/index.js
@@ -1,5 +1,5 @@
 export { default as displayPrice } from './displayPrice';
 export { default as truncate } from './truncate';
-export { shippingOptions, stateOptions, countryOptions } from './orderInfo';
+export { shippingOptions, stateOptions, countryOptions } from './checkout';
 
 export { numSort, alphaSort } from './sort';
